Integration tests for erlang:binary_part/3 including badarg
