+++ native_implemented/otp/src/test/strategy.rs
@@ -10,8 +10,6 @@ use liblumen_alloc::erts::process::Process;
 use liblumen_alloc::erts::term::prelude::*;
 use liblumen_alloc::erts::time::Milliseconds;
 
-use crate::test::total_byte_len;
-
 pub mod base;
 pub mod byte_vec;
 pub mod module_function_arity;
@@ -77,122 +75,6 @@ pub fn term(arc_process: Arc<Process>) -> BoxedStrategy<Term> {
             },
         )
         .boxed()
-}
-
-pub fn with_negative_start_with_valid_length(
-    arc_process: Arc<Process>,
-) -> impl Strategy<Value = (Arc<Process>, Term, Term, Term)> {
-    (
-        Just(arc_process.clone()),
-        term::is_bitstring(arc_process.clone()),
-        term::integer::small::negative(arc_process.clone()),
-    )
-        .prop_flat_map(|(arc_process, binary, start)| {
-            (
-                Just(arc_process.clone()),
-                Just(binary),
-                Just(start),
-                (Just(arc_process.clone()), 0..=total_byte_len(binary))
-                    .prop_map(|(arc_process, length)| arc_process.integer(length)),
-            )
-        })
-}
-
-pub fn with_non_negative_integer_start_without_integer_length(
-    arc_process: Arc<Process>,
-) -> (
-    impl Strategy<Value = Arc<Process>>,
-    impl Strategy<Value = Term>,
-    impl Strategy<Value = Term>,
-    impl Strategy<Value = Term>,
-) {
-    (
-        Just(arc_process.clone()),
-        term::is_bitstring(arc_process.clone()),
-        term::integer::non_negative(arc_process.clone()),
-        term::is_not_integer(arc_process.clone()),
-    )
-}
-
-pub fn with_start_greater_than_size_with_non_negative_length(
-    arc_process: Arc<Process>,
-) -> impl Strategy<Value = (Arc<Process>, Term, Term, Term)> {
-    (
-        Just(arc_process.clone()),
-        term::is_bitstring(arc_process.clone()),
-    )
-        .prop_flat_map(|(arc_process, binary)| {
-            (
-                Just(arc_process.clone()),
-                Just(binary),
-                Just(arc_process.integer(total_byte_len(binary) + 1)),
-                (Just(arc_process.clone()), 0..=total_byte_len(binary))
-                    .prop_map(|(arc_process, length)| arc_process.integer(length)),
-            )
-        })
-}
-
-pub fn with_start_less_than_size_with_negative_length_past_start(
-    arc_process: Arc<Process>,
-) -> impl Strategy<Value = (Arc<Process>, Term, Term, Term, isize)> {
-    (
-        Just(arc_process.clone()),
-        term::is_bitstring::with_byte_len_range(
-            NON_EMPTY_RANGE_INCLUSIVE.into(),
-            arc_process.clone(),
-        ),
-    )
-        .prop_flat_map(|(arc_process, binary)| {
-            (
-                Just(arc_process.clone()),
-                Just(binary),
-                0..total_byte_len(binary),
-            )
-        })
-        .prop_map(|(arc_process, binary, start)| {
-            let length = -((start as isize) + 1);
-            let end = (start as isize) + length;
-
-            (
-                arc_process.clone(),
-                binary,
-                arc_process.integer(start),
-                arc_process.integer(length),
-                end,
-            )
-        })
-}
-
-pub fn without_integer_start_with_integer_length(
-    arc_process: Arc<Process>,
-) -> (
-    impl Strategy<Value = Arc<Process>>,
-    impl Strategy<Value = Term>,
-    impl Strategy<Value = Term>,
-    impl Strategy<Value = Term>,
-) {
-    (
-        Just(arc_process.clone()),
-        term::is_bitstring(arc_process.clone()),
-        term::is_not_integer(arc_process.clone()),
-        term::is_integer(arc_process.clone()),
-    )
-}
-
-pub fn without_integer_start_without_integer_length(
-    arc_process: Arc<Process>,
-) -> (
-    impl Strategy<Value = Arc<Process>>,
-    impl Strategy<Value = Term>,
-    impl Strategy<Value = Term>,
-    impl Strategy<Value = Term>,
-) {
-    (
-        Just(arc_process.clone()),
-        term::is_bitstring(arc_process.clone()),
-        term::is_not_integer(arc_process.clone()),
-        term::is_not_integer(arc_process.clone()),
-    )
 }
 
 const DEPTH: u32 = 3;
