+++ native_implemented/otp/src/test/strategy/term.rs
@@ -20,7 +20,6 @@ pub mod binary;
 pub mod function;
 pub mod index;
 pub mod integer;
-pub mod is_bitstring;
 pub mod list;
 pub mod map;
 pub mod pid;
