--- native_implemented/otp/src/test/strategy/term/is_bitstring.rs
@@ -1,28 +1,0 @@
-use std::ops::RangeInclusive;
-use std::sync::Arc;
-
-use proptest::prop_oneof;
-use proptest::strategy::{BoxedStrategy, Strategy};
-
-use liblumen_alloc::erts::term::prelude::Term;
-use liblumen_alloc::erts::Process;
-
-use super::binary::sub::{bit_count, bit_offset, byte_offset};
-use super::binary::{heap, sub};
-
-pub fn with_byte_len_range(
-    byte_len_range_inclusive: RangeInclusive<usize>,
-    arc_process: Arc<Process>,
-) -> BoxedStrategy<Term> {
-    prop_oneof![
-        heap::with_size_range(byte_len_range_inclusive.clone().into(), arc_process.clone()),
-        sub::with_size_range(
-            byte_offset(),
-            bit_offset(),
-            byte_len_range_inclusive.boxed(),
-            bit_count(),
-            arc_process
-        )
-    ]
-    .boxed()
-}
