+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3.rs
@@ -1,4 +1,4 @@
 #[path = "binary_part_3/with_bitstring.rs"]
 pub mod with_bitstring;
 
-// `without_bitstring_errors_badarg` in unit tests
+test_stdout!(without_bitstring_errors_badarg, "{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n{caught, error, badarg}\n");
