+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3/with_bitstring/with_negative_start_with_valid_length_errors_badarg/init.erl
@@ -1,0 +1,11 @@
+-module(init).
+-export([start/0]).
+-import(erlang, [display/1]).
+
+start() ->
+  Start = -1,
+  display(Start < 0),
+  Length = 0,
+  test:caught(fun () ->
+    binary_part(<<>>, Start, Length)
+  end).
