+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3/with_bitstring/with_start_greater_than_size_with_non_negative_length_errors_badarg/init.erl
@@ -1,0 +1,12 @@
+-module(init).
+-export([start/0]).
+-import(erlang, [display/1]).
+
+start() ->
+  Binary = <<>>,
+  Length = 0,
+  Start = byte_size(Binary) + 1,
+  display(Start > byte_size(Binary)),
+  test:caught(fun () ->
+    binary_part(Binary, Start, Length)
+  end).
