+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3/with_bitstring/with_start_less_than_size_with_positive_length_past_end_errors_badarg/init.erl
@@ -1,0 +1,13 @@
+-module(init).
+-export([start/0]).
+-import(erlang, [display/1]).
+
+start() ->
+  Binary = <<0, 1>>,
+  Start = 1,
+  display(Start < byte_size(Binary)),
+  Length = 2,
+  display(Start + Length =< byte_size(Binary)),
+  test:caught(fun () ->
+    binary_part(Binary, Start, Length)
+  end).
