+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3/with_bitstring/without_integer_start_with_integer_length_errors_badarg/init.erl
@@ -1,0 +1,14 @@
+-module(init).
+-export([start/0]).
+
+start() ->
+  test:each(fun
+    (Integer) when is_integer(Integer) -> ignore;
+    (Term) -> test(Term)
+  end).
+
+test(Start) ->
+  Length = 0,
+  test:caught(fun () ->
+    binary_part(<<>>, Start, Length)
+  end).
