+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3/with_bitstring/without_integer_start_without_integer_length_errors_badarg/init.erl
@@ -1,0 +1,15 @@
+-module(init).
+-export([start/0]).
+
+start() ->
+  test:each(fun
+    (Integer) when is_integer(Integer) -> ignore;
+    (Term) -> test(Term)
+  end).
+
+test(StartOrLength) ->
+  Start = StartOrLength,
+  Length = StartOrLength,
+  test:caught(fun () ->
+    binary_part(<<>>, Start, Length)
+  end).
