+++ native_implemented/otp/tests/internal/lib/erlang/binary_part_3/without_bitstring_errors_badarg/init.erl
@@ -1,0 +1,15 @@
+-module(init).
+-export([start/0]).
+
+start() ->
+  test:each(fun
+    (Bitstring) when is_bitstring(Bitstring) -> ignore;
+    (Term) -> test(Term)
+  end).
+
+test(Binary) ->
+  Start = 0,
+  Length = 0,
+  test:caught(fun () ->
+    binary_part(Binary, Start, Length)
+  end).
